Third pass refactoring, cleanup, and testing.
