+++ geomesa-spark/geomesa-spark-jts/pom.xml
@@ -48,10 +48,6 @@
             <artifactId>specs2_${scala.binary.version}</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>com.typesafe.scala-logging</groupId>
-            <artifactId>scala-logging_${scala.binary.version}</artifactId>
-        </dependency>
     </dependencies>
 
     <build>
