+++ geomesa-spark/geomesa-spark-jts/src/test/scala/org/locationtech/geomesa/spark/jts/udf/GeometricAccessorFunctionsTest.scala
@@ -6,11 +6,12 @@
  * http://www.opensource.org/licenses/apache2.0.php.
  ***********************************************************************/
 
-package org.locationtech.geomesa.spark.jts
+package org.locationtech.geomesa.spark.jts.udf
 
 import com.vividsolutions.jts.geom.Geometry
 import org.apache.spark.sql.functions._
 import org.junit.runner.RunWith
+import org.locationtech.geomesa.spark.jts._
 import org.locationtech.geomesa.spark.jts.util.WKTUtils
 import org.specs2.mutable.Specification
 import org.specs2.runner.JUnitRunner
