+++ geomesa-spark/geomesa-spark-jts/src/test/scala/org/locationtech/geomesa/spark/jts/udf/GeometricOutputFunctionsTest.scala
@@ -6,12 +6,10 @@
  * http://www.opensource.org/licenses/apache2.0.php.
  ***********************************************************************/
 
-package org.locationtech.geomesa.spark.jts
+package org.locationtech.geomesa.spark.jts.udf
 
 import com.vividsolutions.jts.geom.{Coordinate, GeometryFactory}
-import org.apache.spark.sql._
 import org.apache.spark.sql.functions._
-import org.apache.spark.sql.jts.JTSTypes
 import org.junit.runner.RunWith
 import org.locationtech.geomesa.spark.jts._
 import org.specs2.mutable.Specification
